Attribute Macro proconio::derive_readable

#[derive_readable]

Derives Readable for your own type.

If every member of your struct implements Readable, your own type can also be Readable. All you have to do is just add #[derive_readable] to your type definition. This macro automatically implements Readable to your struct and translate your struct’s member type to the output type of the read. For example, if you have Usize1 in your struct, it will actually be defined as usize. Of course the Usize1’s Readable implementation is used to read.