Trait proconio::source::Readable

pub trait Readable {
    type Output;

    // Required method
    fn read<R: BufRead, S: Source<R>>(source: &mut S) -> Self::Output;
}

A trait representing which type can be read from Source.

If you want to read your own type using input!, you can implement this trait for your type. Alternatively, you can add #[derive_readable] if you put use proconio_derive::derive_readable in your source. It automatically implements Readable if all members of your type are Readable.

Required Associated Types

type Output

Required Methods

fn read<R: BufRead, S: Source<R>>(source: &mut S) -> Self::Output

Object Safety

This trait is not object safe.